refactor(backend): Allow guests to manage their shopping carts
-Allow Cart(Order) Creation Without a Usera

-Allow Items to be Added to a Guest Cart: Once the
  guest cart exists, the guest needs to be able to
  add/update/remove items (OrderItems) in it.
  This allows items to be managed in any cart
  that is still "pending", which is perfect for
  guests.

-Prevent Tampering with Completed Orders: While opening
  up access for guests, prevent anyone
  (including guests) from changing an order that has
  already been submitted.

-Restricting listing all orders and orderItems
  to be admin-only. This prevents a malicious user
  from trying to view all the carts in the system.

-Adding a separate, secure endpoint for admins to
  view any specific order they need to.

-Changes the totalamount from an INTEGER to a
  DECIMAL(10, 2) to handle cents.

-Re-orders the dropTable commands in the down function to drop the
  table that contains the foreign key before dropping the table it
  points to.

+++ backend/models/order.js
@@ -1,7 +1,7 @@
 import { Model, DataTypes } from "sequelize";
 import { sequelize } from "../util/db.js";
 
-class Order extends Model {}
+class Order extends Model { }
 
 Order.init(
   {
@@ -11,7 +11,7 @@ Order.init(
       autoIncrement: true,
     },
     totalamount: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(10, 2),
       allowNull: false,
       defaultValue: 0,
     },
